fix(api): change route for post, patch and delete methods
